backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
langchain-openai>=0.1.0
langchain>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0

